refactor: moved Main from __main__ to main
